requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
sqlalchemy
psycopg2-binary
notion-client